requirements.txt
ipykernel==7.1.0
numpy==2.3.5
pandas==2.3.3
biopython==1.86
openpyxl==3.1.5
scikit-learn==1.7.2
joblib>=1.3.0
xgboost>=2.0.0
tensorflow>=2.15.0
pyyaml>=6.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9